fix(vulkan): cast layerName char to u8 before ascii_str() (#256)

+++ vulkan/validation.v
@@ -29,7 +29,7 @@ fn is_validation_layer_available() bool {
 			if c == 0 {
 				break
 			}
-			name += c.ascii_str()
+			name += u8(c).ascii_str()
 		}
 		if name == target {
 			return true
